+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - wifiScanJSON() rewritten, using asyn method WiFi.scanNetworksAsync(wifiScanResult)
 - firmwareOnlineUpdate() updated
 - code CleanUP
+- Fixed Interval For restart device if in AP mode and pass & Gateway are set in config.json | default 5min
 
 **000.05.103**
 - added support for DHT sensors DHT11, DHT12, DHT21, DHT22